[feature/#12-shape-wobble-effect] Three.js 관련 컴포넌트를 별도로 분리합니다
