+++ package.json
@@ -74,6 +74,7 @@
 		"@types/node": "^18.15.12",
 		"@types/react": "^18.0.37",
 		"@types/react-dom": "^18.0.11",
+		"@types/three": "^0.150.2",
 		"@typescript-eslint/eslint-plugin": "^5.58.0",
 		"@typescript-eslint/parser": "^5.58.0",
 		"eslint": "^8.38.0",
